Reorganize HelcityFormalism 	- Move Dynamic functions to an extra module 	- Remove some unnecessary dependencies
