+++ Core/DataPoint.hpp
@@ -77,8 +77,9 @@ public:
       return ret;
     if (n >= Kinematics::Instance()->GetNVars())
       throw std::runtime_error("dataPoint::getRow() | out of range!");
-    for (int i = 0; i < v.size(); i++)
+    for (int i = 0; i < v.size(); i++) {
       ret.push_back(v.at(i).GetValue(n));
+      }
     return ret;
   }
