+++ Physics/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory( DecayDynamics )
 add_subdirectory( HelicityFormalism )
 add_subdirectory( qft++ )
 #add_subdirectory( EvtGen )
