+++ Tools/RootGenerator.hpp
@@ -43,41 +43,42 @@ class RootGenerator : public Generator {
 public:
   //! Constructor for a three particle decay with given masses
   RootGenerator(double sqrtS, double m1, double m2, double m3, int seed = -1);
-  
+
   //! Default Constructor. Information on the decay is obtained from Kinematics
-  RootGenerator(int seed = -1);
-  
+  RootGenerator(std::shared_ptr<Kinematics> kin, int seed = -1);
+
   ~RootGenerator() { delete[] masses; };
 
   virtual RootGenerator *Clone();
-  
+
   virtual void Generate(Event &evt);
-  
+
   virtual void SetSeed(unsigned int seed);
-  
+
   virtual unsigned int GetSeed() const;
-  
+
   virtual double GetUniform(double min, double max) const;
-  
+
   virtual double GetGaussDist(double mu, double sigma) const;
-  
+
   virtual TGenPhaseSpace *GetGenerator() { return &event; }
 
 protected:
   double sqrtS;
-  
+
   TGenPhaseSpace event;
-  
+
   size_t nPart;
-  
+
   Double_t *masses;
 };
 
 class UniformTwoBodyGenerator : public RootGenerator {
 public:
-  UniformTwoBodyGenerator(double minSq_, double maxSq_, int seed = -1)
-      : RootGenerator(seed), minSq(minSq_), maxSq(maxSq_) {
-    if (Kinematics::Instance()->GetFinalState().size() != 2)
+  UniformTwoBodyGenerator(std::shared_ptr<Kinematics> kin, int seed,
+                          double minSq_, double maxSq_)
+      : RootGenerator(kin, seed), minSq(minSq_), maxSq(maxSq_) {
+    if (kin->GetFinalState().size() != 2)
       throw std::runtime_error("UniformTwoBodyGenerator::"
                                "UniformTwoBodyGenerator() | Not a two body "
                                "decay!");
@@ -91,7 +92,7 @@ protected:
   double minSq, maxSq;
 };
 
-} /* namespace Tools*/
+} /* namespace Tools */
 } /* namespace ComPWA */
 
 #endif /* TOOLS_ROOTGENERATOR_HPP_ */
